+++ springboot-practice/.idea/misc.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ExternalStorageConfigurationManager" enabled="true" />
+  <component name="FrameworkDetectionExcludesConfiguration">
+    <file type="web" url="file://$PROJECT_DIR$/jefeel" />
+    <file type="web" url="file://$PROJECT_DIR$/jhlee9462/SpringBootStudy2023" />
+    <file type="web" url="file://$PROJECT_DIR$/joy0987" />
+    <file type="web" url="file://$PROJECT_DIR$/lkdcode" />
+  </component>
+  <component name="ProjectRootManager" version="2" project-jdk-name="corretto-17" project-jdk-type="JavaSDK">
+    <output url="file://$PROJECT_DIR$/out" />
+  </component>
+</project>
