+++ springboot-practice/.idea/modules.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/SpringBootStudy2023.main.iml" filepath="$PROJECT_DIR$/.idea/modules/SpringBootStudy2023.main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/springboot-practice.iml" filepath="$PROJECT_DIR$/.idea/springboot-practice.iml" />
+    </modules>
+  </component>
+</project>
