+++ springboot-practice/jhlee9462/SpringBootStudy2023/src/main/java/com/example/springbootstudy2023/account/repository/AccountRepository.java
@@ -1,0 +1,7 @@
+package com.example.springbootstudy2023.account.repository;
+
+import com.example.springbootstudy2023.account.entity.Account;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccountRepository extends JpaRepository<Account, Long> {
+}
