+++ springboot-practice/jhlee9462/SpringBootStudy2023/src/main/java/com/example/springbootstudy2023/account/service/AccountService.java
@@ -1,0 +1,9 @@
+package com.example.springbootstudy2023.account.service;
+
+import com.example.springbootstudy2023.account.dto.AccountDTO;
+
+public interface AccountService {
+
+    AccountDTO.CreateResponse signUp(AccountDTO.CreateRequest dto);
+
+}
