--- springboot-practice/jhlee9462/SpringBootStudy2023/src/main/java/com/example/springbootstudy2023/exception/ValidateException.java
@@ -1,8 +1,0 @@
-package com.example.springbootstudy2023.exception;
-
-public class ValidateException extends RuntimeException {
-
-    public ValidateException(String message) {
-        super(message);
-    }
-}
