+++ springboot-practice/jhlee9462/SpringBootStudy2023/src/main/java/com/example/springbootstudy2023/global/dto/SearchDTO.java
@@ -1,0 +1,13 @@
+package com.example.springbootstudy2023.global.dto;
+
+public interface SearchDTO {
+
+    record BoardSearchDTO (
+            String keyword
+    ) implements SearchDTO {
+        public static BoardSearchDTO of(String keyword) {
+            return new BoardSearchDTO(keyword);
+        }
+    }
+
+}
