+++ springboot-practice/jhlee9462/SpringBootStudy2023/src/main/java/com/example/springbootstudy2023/global/entity/CommonEntity.java
@@ -1,0 +1,17 @@
+package com.example.springbootstudy2023.global.entity;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
+
+@Getter
+@NoArgsConstructor
+public class CommonEntity {
+    @CreationTimestamp
+    private LocalDateTime createTime;
+    @UpdateTimestamp
+    private LocalDateTime updateTime;
+}
