+++ springboot-practice/jhlee9462/SpringBootStudy2023/src/main/java/com/example/springbootstudy2023/global/exception/ValidateException.java
@@ -1,0 +1,8 @@
+package com.example.springbootstudy2023.global.exception;
+
+public class ValidateException extends RuntimeException {
+
+    public ValidateException(String message) {
+        super(message);
+    }
+}
