+++ springboot-practice/jhlee9462/SpringBootStudy2023/src/main/java/com/example/springbootstudy2023/global/utils/Validator.java
@@ -1,0 +1,9 @@
+package com.example.springbootstudy2023.global.utils;
+
+import com.example.springbootstudy2023.global.exception.ValidateException;
+
+public interface Validator {
+
+    void validate(Object target) throws ValidateException;
+
+}
