--- springboot-practice/jhlee9462/SpringBootStudy2023/src/main/java/com/example/springbootstudy2023/utils/Validator.java
@@ -1,9 +1,0 @@
-package com.example.springbootstudy2023.utils;
-
-import com.example.springbootstudy2023.exception.ValidateException;
-
-public interface Validator {
-
-    void validate(Object target) throws ValidateException;
-
-}
