+++ springboot-practice/jhlee9462/SpringBootStudy2023/src/main/resources/application.yml
@@ -26,3 +26,8 @@ spring:
 # Logging Configuration
 logging.level:
   org.hibernate.SQL: debug
+
+# Jwt Configuration
+jwt:
+  secret: jhlee9462
+  expire: 3600
